feat: Update ScrollBar to winui3/release/1.6.5
